Generate Devtools bindings for specific Chromium version

+++ crawler/entrypoint.sh
@@ -1,5 +1,9 @@
-cp /vpn_shared/resolv.conf /etc/resolv.conf
-python /opt/priv-accept/priv-accept.py \
+#!/bin/sh
+if [ -f "/vpn_shared/resolv.conf" ]; then
+    cp /vpn_shared/resolv.conf /etc/resolv.conf
+fi
+
+python3 /opt/priv-accept/priv-accept.py \
     --chrome_driver /opt/chromium-topics/chromedriver \
     --chrome_binary /opt/chromium-topics/chrome \
     --docker --detect_topics \
